refactor(LottoAplicationController): 책임에 따라 controller 분리
1. LottoAplicationController를 LottoGame을 관리하는 모체 클래스로 변경

2. BuyLottoController(로또 구매 ~ 확인)을 자식 클래스로 추가

3. 변경된 controller에 따른 step3-index.ts 변경

4. new Instance() 형태가 불가능함에 따른 eslint rules 추가(no-new를 off로 변경)
