+++ src/step3/view/index.ts
@@ -1,3 +1,4 @@
 export { default as View } from './View';
 export { default as LottoPurchaseFormView } from './LottoPurchaseFormView';
 export { default as PurchasedLottoView } from './PurchasedLottoView';
+export { default as WinningLottoInfoFormView } from './WinningLottoInfoFormView';
